More NOVA port work

+++ src/main/scala/com/resonant/core/prefab/block/multiblock/IMultiBlock.java
@@ -1,15 +1,16 @@
 package com.resonant.core.prefab.block.multiblock;
 
-import nova.core.util.transform.Vector3d;
+import nova.core.util.transform.Vector3i;
 
 /**
- * Interface to be applied to tile entity blocks that occupies more than one block space. Useful for
+ * Interface to be applied to block entity blocks that occupies more than one block space. Useful for
  * large machines.
+ *
  * @author Calclavia
  */
 public interface IMultiBlock {
 	/**
 	 * @return An array of Vector3d containing the multiblock relative coordinates to be constructed.
 	 */
-	public Iterable<Vector3d> getMultiBlockVectors();
+	public Iterable<Vector3i> getMultiBlockVectors();
 }
